Archive all crunching files in RIDL

RIDL is UNHCR instance of a CKAN server and is accessible for UNHCR staff at https://ridl.unhcr.org . It is designed to keep track and document dataset within an organisation.

Usage

add_ridl(ridl, datafolder, cleaned_data, cleaning_log, namethisfile)

Arguments

ridl

ridl container where the resources should be added

datafolder

folder where the data used by the notebook are stored

cleaned_data

names of the file with the cleaned data

cleaning_log

names of the file with the cleaning_log

namethisfile

all files are archived based on the name of notebook you created. The function automatically get the name of the notebook where it is run from, using basename(rstudioapi::getSourceEditorContext()$path )

Value

nothing all analysis files are added as a resources

Details

You conveniently archive there your generated cleaning report and save the work you did on a notebook: As you have been working on the data, you want to keep track of it and save your work in a place where it can be useful for other people and available for peer review and quality assessment.

The function saves within the the RIDL container you used to get the data from the following resources:

the source notebook

the cleaning log

the cleaned data

The function behavior is the following -

Get metadata from the RIDL dataset

check if the resources to be uploaded is already shared based on the name

if already there update, if not create

The function relies on # install.packages("pak") # pak::pkg_install("edouard-legoupil/riddle")

Examples

if (FALSE) { # \dontrun{
### Example used for each template
## Time to archive your work once done!!
# namethisfile = basename(rstudioapi::getSourceEditorContext()$path )
# if( params$publish == "yes"){
#   add_ridl(ridl = params$ridl,
#             datafolder = params$datafolder,
#             form = params$form,
#             namethisfile =  namethisfile ) }

} # }