Detect select multiple parent columns

auto_detect_sm_parents is mean to detect select multiple parent columns in a way that does not rely on the XLSForm as the input

Usage

auto_detect_sm_parents(dataset, sm_separator = ".")

Arguments

dataset

dataset to correct

sm_separator

Separator for choice multiple questions. The default is "." If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'

Value

a list of select multiple parent columns in data set.