FCS component checks

FCS component checks

Usage

check_fcs(
  dataset,
  uuid_column = "uuid",
  cereals_column = "fs_fcs_cerealgrainroottuber",
  pulses_column = "fs_fcs_beansnuts",
  dairy_column = "fs_fcs_dairy",
  meat_column = "fs_fcs_meatfishegg",
  vegetables_column = "fs_fcs_vegetableleave",
  fruits_column = "fs_fcs_fruit",
  oil_column = "fs_fcs_fat",
  sugar_column = "fs_fcs_sugar"
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

uuid column in the dataset. Default is uuid.

cereals_column

Variable name for- In the last 7 days, on how many days did your household eat cereals, grains, roots and tubers, including wild roots?

pulses_column

Variable name for- In the last 7 days, on how many days did your household eat any beans / legumes, pulses or nuts?

dairy_column

Variable name for- In the last 7 days, on how many days did your household drink milk or eat other dairy products?

meat_column

Variable name for- In the last 7 days, on how many days did your household eat meat, fish or eggs?

vegetables_column

Variable name for- In the last 7 days, on how many days did your household eat vegetables or leaves, including all wild vegetables and leaves?

fruits_column

Variable name for- In the last 7 days, on how many days did your household eat fruits, including all wild fruits?

oil_column

Variable name for- In the last 7 days, on how many days did your household eat oil, fat, or butter?

sugar_column

Variable name for- In the last 7 days, on how many days did your household eat sugar or sugary foods?

Value

dataset with potential issues