check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        46     285     60 100
#> 2     uuid_2        49     499      0 100
#> 3     uuid_3        54     293     80 100
#> 4     uuid_4        52     246   1020 100
#> 5     uuid_5        50     349   1050 100
#> 6     uuid_6        47     482  38055 100
#> 7     uuid_7        47     391  33012 100
#> 8     uuid_8        53     405  33624 100
#> 9     uuid_9        45     449  30380 100
#> 10   uuid_10        53     353  39104 100
#> 11   uuid_11        51     290  41107 100
#> 12   uuid_12        47     386  35429 100
#> 13   uuid_13        52     236  28994 100
#> 14   uuid_14        51     461  32031 100
#> 15   uuid_15        47     259  47516 100
#> 16   uuid_16        53     483  47437 100
#> 17   uuid_17        54     463  49987 100
#> 18   uuid_18        51     301  21379 100
#> 19   uuid_19        52     486  24110 100
#> 20   uuid_20        48     237  36440 100
#> 21   uuid_21        52     436  22935 100
#> 22   uuid_22        45     254  35113 100
#> 23   uuid_23        48     355  40958 100
#> 24   uuid_24        52     359  45660 100
#> 25   uuid_25        50     339  34955 100
#> 26   uuid_26        45     279  33842 100
#> 27   uuid_27        53     287  47951 100
#> 28   uuid_28        54     371  34181 100
#> 29   uuid_29        53     268  45634 100
#> 30   uuid_30        47     496  46471 100
#> 31   uuid_31        52     419  45896 100
#> 32   uuid_32        51     232  33275 100
#> 33   uuid_33        50     374  35803 100
#> 34   uuid_34        53     212  40724 100
#> 35   uuid_35        49     211  26184 100
#> 36   uuid_36        54     346  27378 100
#> 37   uuid_37        53     212  27009 100
#> 38   uuid_38        47     209  36606 100
#> 39   uuid_39        53     453  44706 100
#> 40   uuid_40        48     437  38526 100
#> 41   uuid_41        51     264  48160 100
#> 42   uuid_42        47     246  29559 100
#> 43   uuid_43        45     270  37788 100
#> 44   uuid_44        48     230  43952 100
#> 45   uuid_45        54     259  26996 100
#> 46   uuid_46        46     478  21870 100
#> 47   uuid_47        49     325  35017 100
#> 48   uuid_48        48     492  38443 100
#> 49   uuid_49        48     471  24086 100
#> 50   uuid_50        46     214  41133 100
#> 51   uuid_51        49     443  38724 100
#> 52   uuid_52        47     482  49637 100
#> 53   uuid_53        51     457  44555 100
#> 54   uuid_54        50     314  24714 100
#> 55   uuid_55        54     380  44822 100
#> 56   uuid_56        51     489  36492 100
#> 57   uuid_57        51     344  20299 100
#> 58   uuid_58        47     289  21020 100
#> 59   uuid_59        54     466  25629 100
#> 60   uuid_60        46     341  24066 100
#> 61   uuid_61        53     480  38073 100
#> 62   uuid_62        54     318  45626 100
#> 63   uuid_63        45     301  45525 100
#> 64   uuid_64        52     214  38143 100
#> 65   uuid_65        49     219  36279 100
#> 66   uuid_66        48     323  47266 100
#> 67   uuid_67        48     363  34000 100
#> 68   uuid_68        47     421  27757 100
#> 69   uuid_69        52     393  20612 100
#> 70   uuid_70        54     381  31024 100
#> 71   uuid_71        52     307  38256 100
#> 72   uuid_72        53     280  46706 100
#> 73   uuid_73        54     449  23359 100
#> 74   uuid_74        51     227  49222 100
#> 75   uuid_75        49     381  42631 100
#> 76   uuid_76        51     423  46185 100
#> 77   uuid_77        48     305  39561 100
#> 78   uuid_78        52     402  39675 100
#> 79   uuid_79        53     329  27166 100
#> 80   uuid_80        50     328  29885 100
#> 81   uuid_81        55     278  46070 100
#> 82   uuid_82        47     429  39540 100
#> 83   uuid_83        45     216  41350 100
#> 84   uuid_84        51     410  29428 100
#> 85   uuid_85        47     338  32301 100
#> 86   uuid_86        47     258  26025 100
#> 87   uuid_87        53     397  35968 100
#> 88   uuid_88        52     351  30687 100
#> 89   uuid_89        53     275  36734 100
#> 90   uuid_90        46     212  26884 100
#> 91   uuid_91         1     243  39644 100
#> 92   uuid_92         1     355  28473 100
#> 93   uuid_93         1     343  28216 100
#> 94   uuid_94         1     238  21487 100
#> 95   uuid_95         1     307  43162 100
#> 96   uuid_96        99     600  46968 100
#> 97   uuid_97        99     100  40367 100
#> 98   uuid_98        99      80  44140 100
#> 99   uuid_99       100    1020  29554 100
#> 100 uuid_100        99    1050  26239  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (normal distribution) income           60
#> 16 uuid_2   outlier (normal distribution) income            0
#> 17 uuid_3   outlier (normal distribution) income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>