cleaningtools: cleaningtools package focuses on data cleaning

The cleaningtools package focuses on cleaning, and has three components: **Check**, which includes a set of functions that flag values, such as check_outliers and check_logical. **Create**, which includes a set of functions to create different items for use in cleaning, such as the cleaning log from the checks, clean data, and enumerator performance. **Review**, which includes a set of functions to review the cleaning.

See also

Author

Maintainer: Yann Say yann.say@impact-initiatives.org (ORCID)

Authors:

Mehedi Khan mh.khan@reach-initiative.org (ORCID)