Review cleaning log

Review the cleaning log. It can be used before create_clean_data to check cleaning log. Possible flags are:

Value in change_response_value not found

question_does_not_exist: question does not exist

uuid_does_not_exist: cannot find a UUID

na_in_change_type: change_type is NA

Usage

review_cleaning_log(
  raw_dataset,
  raw_data_uuid_column = "uuid",
  cleaning_log,
  cleaning_log_uuid_column = "uuid",
  cleaning_log_question_column,
  cleaning_log_new_value_column,
  cleaning_log_change_type_column,
  change_response_value = "change_response"
)

Arguments

raw_dataset

Raw dataset

raw_data_uuid_column

uuid column in the raw dataset. Default is "uuid".

cleaning_log

cleaning log (data.frame)

cleaning_log_uuid_column

uuid column in the cleaning log. Default is "uuid".

cleaning_log_question_column

column in cleaning log which specifies data set column to change

cleaning_log_new_value_column

cleaning log column specifying the new correct value

cleaning_log_change_type_column

column in cleaning log which specifies change type to be made

change_response_value

values in change type column which should be changed to a new value.

Value

review of the cleaning log with flagged entries and note explaining the flags.