check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), 
                round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, 
                     replace = TRUE, size = 95), 
              c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), 
             sample(20000:50000, 
                    replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier,
               uuid_column = "uuid") |>
  knitr::kable()
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> 
#> 
#> |uuid     | one_value| expense| income|  yy|
#> |:--------|---------:|-------:|------:|---:|
#> |uuid_1   |        46|     290|     60| 100|
#> |uuid_2   |        51|     322|      0| 100|
#> |uuid_3   |        49|     483|     80| 100|
#> |uuid_4   |        50|     332|   1020| 100|
#> |uuid_5   |        46|     210|   1050| 100|
#> |uuid_6   |        51|     328|  34973| 100|
#> |uuid_7   |        45|     358|  31594| 100|
#> |uuid_8   |        55|     275|  47372| 100|
#> |uuid_9   |        51|     249|  27895| 100|
#> |uuid_10  |        48|     366|  38683| 100|
#> |uuid_11  |        54|     222|  39726| 100|
#> |uuid_12  |        48|     457|  26535| 100|
#> |uuid_13  |        51|     229|  46745| 100|
#> |uuid_14  |        46|     376|  32384| 100|
#> |uuid_15  |        46|     457|  33422| 100|
#> |uuid_16  |        45|     213|  31659| 100|
#> |uuid_17  |        46|     245|  20189| 100|
#> |uuid_18  |        47|     288|  35672| 100|
#> |uuid_19  |        46|     304|  24211| 100|
#> |uuid_20  |        53|     393|  41929| 100|
#> |uuid_21  |        50|     253|  42530| 100|
#> |uuid_22  |        47|     254|  25055| 100|
#> |uuid_23  |        48|     492|  24058| 100|
#> |uuid_24  |        54|     483|  39046| 100|
#> |uuid_25  |        46|     426|  34686| 100|
#> |uuid_26  |        48|     437|  49454| 100|
#> |uuid_27  |        47|     479|  32365| 100|
#> |uuid_28  |        55|     485|  46970| 100|
#> |uuid_29  |        48|     328|  29405| 100|
#> |uuid_30  |        53|     467|  28443| 100|
#> |uuid_31  |        50|     424|  47511| 100|
#> |uuid_32  |        53|     370|  28968| 100|
#> |uuid_33  |        54|     265|  29098| 100|
#> |uuid_34  |        48|     274|  34274| 100|
#> |uuid_35  |        49|     401|  29534| 100|
#> |uuid_36  |        46|     448|  41295| 100|
#> |uuid_37  |        48|     346|  21059| 100|
#> |uuid_38  |        51|     304|  46084| 100|
#> |uuid_39  |        48|     445|  28917| 100|
#> |uuid_40  |        53|     269|  26413| 100|
#> |uuid_41  |        48|     464|  32410| 100|
#> |uuid_42  |        47|     353|  43416| 100|
#> |uuid_43  |        49|     391|  20400| 100|
#> |uuid_44  |        48|     229|  35293| 100|
#> |uuid_45  |        46|     361|  47018| 100|
#> |uuid_46  |        53|     455|  49589| 100|
#> |uuid_47  |        47|     457|  38425| 100|
#> |uuid_48  |        53|     406|  32473| 100|
#> |uuid_49  |        54|     396|  25670| 100|
#> |uuid_50  |        45|     394|  31991| 100|
#> |uuid_51  |        54|     484|  23604| 100|
#> |uuid_52  |        53|     368|  46289| 100|
#> |uuid_53  |        47|     434|  44377| 100|
#> |uuid_54  |        46|     365|  30144| 100|
#> |uuid_55  |        46|     335|  25535| 100|
#> |uuid_56  |        49|     368|  36944| 100|
#> |uuid_57  |        54|     337|  45969| 100|
#> |uuid_58  |        47|     314|  20189| 100|
#> |uuid_59  |        50|     202|  26464| 100|
#> |uuid_60  |        46|     359|  29271| 100|
#> |uuid_61  |        51|     422|  24041| 100|
#> |uuid_62  |        55|     392|  40410| 100|
#> |uuid_63  |        52|     500|  38255| 100|
#> |uuid_64  |        53|     240|  39796| 100|
#> |uuid_65  |        51|     474|  25041| 100|
#> |uuid_66  |        53|     387|  20509| 100|
#> |uuid_67  |        53|     288|  26059| 100|
#> |uuid_68  |        48|     459|  20023| 100|
#> |uuid_69  |        55|     245|  20537| 100|
#> |uuid_70  |        46|     418|  31565| 100|
#> |uuid_71  |        50|     405|  26912| 100|
#> |uuid_72  |        51|     276|  27800| 100|
#> |uuid_73  |        47|     340|  49173| 100|
#> |uuid_74  |        49|     490|  33803| 100|
#> |uuid_75  |        50|     361|  20871| 100|
#> |uuid_76  |        52|     238|  35952| 100|
#> |uuid_77  |        46|     334|  22471| 100|
#> |uuid_78  |        50|     200|  39152| 100|
#> |uuid_79  |        52|     262|  34900| 100|
#> |uuid_80  |        46|     329|  43300| 100|
#> |uuid_81  |        50|     280|  44298| 100|
#> |uuid_82  |        51|     498|  42663| 100|
#> |uuid_83  |        51|     275|  33966| 100|
#> |uuid_84  |        54|     412|  30276| 100|
#> |uuid_85  |        49|     455|  48444| 100|
#> |uuid_86  |        51|     280|  20504| 100|
#> |uuid_87  |        51|     221|  31234| 100|
#> |uuid_88  |        50|     305|  24052| 100|
#> |uuid_89  |        54|     299|  26206| 100|
#> |uuid_90  |        53|     375|  34934| 100|
#> |uuid_91  |         0|     267|  22116| 100|
#> |uuid_92  |         0|     415|  26038| 100|
#> |uuid_93  |         1|     340|  27966| 100|
#> |uuid_94  |         1|     208|  22658| 100|
#> |uuid_95  |         0|     322|  46817| 100|
#> |uuid_96  |       100|     600|  45013| 100|
#> |uuid_97  |        99|     100|  42582| 100|
#> |uuid_98  |        99|      80|  21883| 100|
#> |uuid_99  |       100|    1020|  35174| 100|
#> |uuid_100 |        99|    1050|  28105|  10|
#> 
#> |uuid     |issue                         |question  | old_value|
#> |:--------|:-----------------------------|:---------|---------:|
#> |uuid_91  |outlier (normal distribution) |one_value |         0|
#> |uuid_92  |outlier (normal distribution) |one_value |         0|
#> |uuid_93  |outlier (normal distribution) |one_value |         1|
#> |uuid_94  |outlier (normal distribution) |one_value |         1|
#> |uuid_95  |outlier (normal distribution) |one_value |         0|
#> |uuid_96  |outlier (normal distribution) |one_value |       100|
#> |uuid_97  |outlier (normal distribution) |one_value |        99|
#> |uuid_98  |outlier (normal distribution) |one_value |        99|
#> |uuid_99  |outlier (normal distribution) |one_value |       100|
#> |uuid_100 |outlier (normal distribution) |one_value |        99|
#> |uuid_99  |outlier (normal distribution) |expense   |      1020|
#> |uuid_100 |outlier (normal distribution) |expense   |      1050|
#> |uuid_97  |outlier (log distribution)    |expense   |       100|
#> |uuid_98  |outlier (log distribution)    |expense   |        80|
#> |uuid_1   |outlier (log distribution)    |income    |        60|
#> |uuid_2   |outlier (log distribution)    |income    |         0|
#> |uuid_3   |outlier (log distribution)    |income    |        80|
#> |uuid_100 |outlier (normal distribution) |yy        |        10|