check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        46     327     60 100
#> 2     uuid_2        52     299      0 100
#> 3     uuid_3        54     478     80 100
#> 4     uuid_4        46     464   1020 100
#> 5     uuid_5        52     264   1050 100
#> 6     uuid_6        54     353  20420 100
#> 7     uuid_7        52     338  23907 100
#> 8     uuid_8        48     296  39322 100
#> 9     uuid_9        50     483  49960 100
#> 10   uuid_10        48     224  39510 100
#> 11   uuid_11        51     281  34776 100
#> 12   uuid_12        53     442  27425 100
#> 13   uuid_13        46     379  31110 100
#> 14   uuid_14        52     280  25635 100
#> 15   uuid_15        54     353  34992 100
#> 16   uuid_16        51     220  22657 100
#> 17   uuid_17        49     405  28311 100
#> 18   uuid_18        48     253  40598 100
#> 19   uuid_19        53     392  48372 100
#> 20   uuid_20        53     266  23170 100
#> 21   uuid_21        46     489  21085 100
#> 22   uuid_22        51     272  31654 100
#> 23   uuid_23        45     397  44969 100
#> 24   uuid_24        52     325  44754 100
#> 25   uuid_25        54     317  43259 100
#> 26   uuid_26        46     464  29232 100
#> 27   uuid_27        54     230  22890 100
#> 28   uuid_28        54     455  43996 100
#> 29   uuid_29        49     392  20581 100
#> 30   uuid_30        46     273  26004 100
#> 31   uuid_31        52     289  44570 100
#> 32   uuid_32        47     313  48345 100
#> 33   uuid_33        54     464  36670 100
#> 34   uuid_34        54     296  42152 100
#> 35   uuid_35        54     326  27997 100
#> 36   uuid_36        52     272  20512 100
#> 37   uuid_37        51     302  28772 100
#> 38   uuid_38        53     464  39335 100
#> 39   uuid_39        46     460  24060 100
#> 40   uuid_40        51     201  45663 100
#> 41   uuid_41        50     310  26477 100
#> 42   uuid_42        46     252  35866 100
#> 43   uuid_43        55     313  22386 100
#> 44   uuid_44        48     241  20799 100
#> 45   uuid_45        46     258  38020 100
#> 46   uuid_46        49     386  49728 100
#> 47   uuid_47        48     462  42193 100
#> 48   uuid_48        51     372  31492 100
#> 49   uuid_49        50     432  37963 100
#> 50   uuid_50        46     436  21040 100
#> 51   uuid_51        53     358  44351 100
#> 52   uuid_52        53     342  30338 100
#> 53   uuid_53        45     405  45496 100
#> 54   uuid_54        53     421  25136 100
#> 55   uuid_55        55     301  41550 100
#> 56   uuid_56        53     486  44505 100
#> 57   uuid_57        46     496  36674 100
#> 58   uuid_58        50     454  44087 100
#> 59   uuid_59        49     333  35275 100
#> 60   uuid_60        47     205  22847 100
#> 61   uuid_61        55     301  47929 100
#> 62   uuid_62        51     381  25170 100
#> 63   uuid_63        49     260  45203 100
#> 64   uuid_64        55     449  31668 100
#> 65   uuid_65        46     325  37096 100
#> 66   uuid_66        49     286  32249 100
#> 67   uuid_67        53     360  38514 100
#> 68   uuid_68        48     335  32873 100
#> 69   uuid_69        50     216  33202 100
#> 70   uuid_70        53     207  47245 100
#> 71   uuid_71        53     203  40749 100
#> 72   uuid_72        52     427  31405 100
#> 73   uuid_73        48     422  42070 100
#> 74   uuid_74        49     282  37436 100
#> 75   uuid_75        50     220  41648 100
#> 76   uuid_76        49     234  28794 100
#> 77   uuid_77        52     424  30955 100
#> 78   uuid_78        49     363  29959 100
#> 79   uuid_79        48     220  35394 100
#> 80   uuid_80        45     362  27266 100
#> 81   uuid_81        55     322  25869 100
#> 82   uuid_82        46     495  33767 100
#> 83   uuid_83        49     436  38499 100
#> 84   uuid_84        48     290  34437 100
#> 85   uuid_85        49     453  38286 100
#> 86   uuid_86        51     282  39722 100
#> 87   uuid_87        53     272  21182 100
#> 88   uuid_88        52     370  40564 100
#> 89   uuid_89        52     250  28860 100
#> 90   uuid_90        52     442  33328 100
#> 91   uuid_91         0     410  48266 100
#> 92   uuid_92         0     327  44018 100
#> 93   uuid_93         1     490  39303 100
#> 94   uuid_94         1     344  30415 100
#> 95   uuid_95         1     311  22666 100
#> 96   uuid_96       100     600  20197 100
#> 97   uuid_97       100     100  48164 100
#> 98   uuid_98       100      80  48865 100
#> 99   uuid_99       100    1020  33127 100
#> 100 uuid_100       100    1050  29067  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         0
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value       100
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>