check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        53     246     60 100
#> 2     uuid_2        48     283      0 100
#> 3     uuid_3        47     243     80 100
#> 4     uuid_4        47     347   1020 100
#> 5     uuid_5        52     480   1050 100
#> 6     uuid_6        49     313  35640 100
#> 7     uuid_7        49     252  43910 100
#> 8     uuid_8        46     321  20482 100
#> 9     uuid_9        50     411  23168 100
#> 10   uuid_10        53     310  35127 100
#> 11   uuid_11        53     338  43135 100
#> 12   uuid_12        53     469  35267 100
#> 13   uuid_13        49     346  47530 100
#> 14   uuid_14        52     406  32577 100
#> 15   uuid_15        49     266  25581 100
#> 16   uuid_16        54     339  33126 100
#> 17   uuid_17        46     297  26012 100
#> 18   uuid_18        45     417  24999 100
#> 19   uuid_19        47     357  26662 100
#> 20   uuid_20        47     386  33346 100
#> 21   uuid_21        54     452  46913 100
#> 22   uuid_22        53     414  21105 100
#> 23   uuid_23        48     368  30428 100
#> 24   uuid_24        50     221  23711 100
#> 25   uuid_25        55     389  23924 100
#> 26   uuid_26        54     450  48593 100
#> 27   uuid_27        47     325  26892 100
#> 28   uuid_28        48     392  33787 100
#> 29   uuid_29        53     275  20964 100
#> 30   uuid_30        54     406  39107 100
#> 31   uuid_31        49     321  29410 100
#> 32   uuid_32        49     405  33280 100
#> 33   uuid_33        51     460  25546 100
#> 34   uuid_34        54     470  38094 100
#> 35   uuid_35        53     295  40675 100
#> 36   uuid_36        48     207  25068 100
#> 37   uuid_37        49     398  45091 100
#> 38   uuid_38        49     474  25330 100
#> 39   uuid_39        47     461  34717 100
#> 40   uuid_40        52     260  37965 100
#> 41   uuid_41        54     490  40336 100
#> 42   uuid_42        46     390  34449 100
#> 43   uuid_43        46     400  20386 100
#> 44   uuid_44        51     427  28339 100
#> 45   uuid_45        49     462  37069 100
#> 46   uuid_46        45     353  47299 100
#> 47   uuid_47        49     409  29534 100
#> 48   uuid_48        53     309  37688 100
#> 49   uuid_49        50     310  35812 100
#> 50   uuid_50        49     484  37608 100
#> 51   uuid_51        51     328  26056 100
#> 52   uuid_52        53     400  30028 100
#> 53   uuid_53        48     234  46105 100
#> 54   uuid_54        46     218  48539 100
#> 55   uuid_55        53     356  23628 100
#> 56   uuid_56        53     232  37043 100
#> 57   uuid_57        52     200  29839 100
#> 58   uuid_58        52     370  28784 100
#> 59   uuid_59        51     299  41147 100
#> 60   uuid_60        45     290  28497 100
#> 61   uuid_61        51     280  31844 100
#> 62   uuid_62        55     312  44159 100
#> 63   uuid_63        49     457  27011 100
#> 64   uuid_64        48     323  32932 100
#> 65   uuid_65        51     328  45846 100
#> 66   uuid_66        54     248  41687 100
#> 67   uuid_67        50     347  49727 100
#> 68   uuid_68        46     340  35870 100
#> 69   uuid_69        50     499  24170 100
#> 70   uuid_70        54     417  47273 100
#> 71   uuid_71        54     219  24629 100
#> 72   uuid_72        55     207  39253 100
#> 73   uuid_73        48     228  25813 100
#> 74   uuid_74        55     472  29209 100
#> 75   uuid_75        48     279  49726 100
#> 76   uuid_76        45     369  35037 100
#> 77   uuid_77        49     275  46543 100
#> 78   uuid_78        49     304  33666 100
#> 79   uuid_79        50     325  26738 100
#> 80   uuid_80        45     497  20533 100
#> 81   uuid_81        45     333  33235 100
#> 82   uuid_82        50     340  43826 100
#> 83   uuid_83        53     396  39195 100
#> 84   uuid_84        49     344  46194 100
#> 85   uuid_85        46     418  39996 100
#> 86   uuid_86        53     377  47592 100
#> 87   uuid_87        55     407  36610 100
#> 88   uuid_88        47     404  30282 100
#> 89   uuid_89        49     445  47569 100
#> 90   uuid_90        46     304  38759 100
#> 91   uuid_91         1     203  28379 100
#> 92   uuid_92         1     468  42899 100
#> 93   uuid_93         0     369  36987 100
#> 94   uuid_94         0     267  30591 100
#> 95   uuid_95         1     482  23872 100
#> 96   uuid_96        99     600  49459 100
#> 97   uuid_97        99     100  28963 100
#> 98   uuid_98       100      80  31760 100
#> 99   uuid_99        99    1020  42393 100
#> 100 uuid_100        99    1050  38172  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>