check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        50     309     60 100
#> 2     uuid_2        50     443      0 100
#> 3     uuid_3        49     457     80 100
#> 4     uuid_4        46     495   1020 100
#> 5     uuid_5        45     326   1050 100
#> 6     uuid_6        50     216  23511 100
#> 7     uuid_7        47     201  20810 100
#> 8     uuid_8        52     267  30611 100
#> 9     uuid_9        50     339  37651 100
#> 10   uuid_10        50     239  45652 100
#> 11   uuid_11        50     334  46494 100
#> 12   uuid_12        48     306  34192 100
#> 13   uuid_13        50     225  37005 100
#> 14   uuid_14        55     475  40131 100
#> 15   uuid_15        50     210  39307 100
#> 16   uuid_16        51     382  41332 100
#> 17   uuid_17        51     336  43793 100
#> 18   uuid_18        54     483  33949 100
#> 19   uuid_19        49     386  24939 100
#> 20   uuid_20        48     301  33347 100
#> 21   uuid_21        50     445  35860 100
#> 22   uuid_22        55     252  31957 100
#> 23   uuid_23        54     302  20532 100
#> 24   uuid_24        55     280  38367 100
#> 25   uuid_25        46     422  24541 100
#> 26   uuid_26        51     327  48470 100
#> 27   uuid_27        47     218  46706 100
#> 28   uuid_28        51     499  22152 100
#> 29   uuid_29        45     208  29401 100
#> 30   uuid_30        48     331  43509 100
#> 31   uuid_31        45     397  29968 100
#> 32   uuid_32        46     214  29872 100
#> 33   uuid_33        53     477  30772 100
#> 34   uuid_34        46     463  28033 100
#> 35   uuid_35        48     464  47613 100
#> 36   uuid_36        47     409  33605 100
#> 37   uuid_37        47     285  31625 100
#> 38   uuid_38        53     299  36540 100
#> 39   uuid_39        53     404  21609 100
#> 40   uuid_40        54     440  48415 100
#> 41   uuid_41        48     255  45518 100
#> 42   uuid_42        53     313  42222 100
#> 43   uuid_43        48     448  24639 100
#> 44   uuid_44        53     340  29749 100
#> 45   uuid_45        51     382  25222 100
#> 46   uuid_46        45     456  48539 100
#> 47   uuid_47        49     466  21447 100
#> 48   uuid_48        52     376  48166 100
#> 49   uuid_49        48     319  27586 100
#> 50   uuid_50        52     456  36577 100
#> 51   uuid_51        51     249  39821 100
#> 52   uuid_52        51     277  38620 100
#> 53   uuid_53        48     483  42989 100
#> 54   uuid_54        46     245  36608 100
#> 55   uuid_55        45     212  33798 100
#> 56   uuid_56        46     317  31513 100
#> 57   uuid_57        53     241  45548 100
#> 58   uuid_58        49     332  49783 100
#> 59   uuid_59        55     316  47512 100
#> 60   uuid_60        47     215  30450 100
#> 61   uuid_61        54     276  43834 100
#> 62   uuid_62        52     217  27905 100
#> 63   uuid_63        54     269  24667 100
#> 64   uuid_64        47     499  38713 100
#> 65   uuid_65        51     455  38051 100
#> 66   uuid_66        54     358  23194 100
#> 67   uuid_67        52     499  40385 100
#> 68   uuid_68        54     294  48564 100
#> 69   uuid_69        50     479  24654 100
#> 70   uuid_70        46     396  30781 100
#> 71   uuid_71        52     343  36314 100
#> 72   uuid_72        47     346  48247 100
#> 73   uuid_73        50     359  27810 100
#> 74   uuid_74        50     371  21303 100
#> 75   uuid_75        52     454  23256 100
#> 76   uuid_76        53     368  20776 100
#> 77   uuid_77        51     443  43610 100
#> 78   uuid_78        53     335  31544 100
#> 79   uuid_79        52     250  28985 100
#> 80   uuid_80        46     369  21667 100
#> 81   uuid_81        53     305  28427 100
#> 82   uuid_82        53     246  23830 100
#> 83   uuid_83        48     366  49679 100
#> 84   uuid_84        50     250  24798 100
#> 85   uuid_85        53     441  31396 100
#> 86   uuid_86        51     295  24383 100
#> 87   uuid_87        50     337  23705 100
#> 88   uuid_88        48     372  32828 100
#> 89   uuid_89        46     282  26315 100
#> 90   uuid_90        55     395  38574 100
#> 91   uuid_91         0     245  40217 100
#> 92   uuid_92         1     342  22430 100
#> 93   uuid_93         1     231  45661 100
#> 94   uuid_94         1     389  39388 100
#> 95   uuid_95         1     396  24190 100
#> 96   uuid_96       100     600  20949 100
#> 97   uuid_97        99     100  42766 100
#> 98   uuid_98       100      80  34729 100
#> 99   uuid_99        99    1020  31150 100
#> 100 uuid_100       100    1050  41604  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>