check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        51     404     60 100
#> 2     uuid_2        53     355      0 100
#> 3     uuid_3        50     483     80 100
#> 4     uuid_4        49     428   1020 100
#> 5     uuid_5        53     261   1050 100
#> 6     uuid_6        51     350  29035 100
#> 7     uuid_7        52     244  26188 100
#> 8     uuid_8        48     494  46289 100
#> 9     uuid_9        48     407  27581 100
#> 10   uuid_10        54     204  34102 100
#> 11   uuid_11        55     203  33941 100
#> 12   uuid_12        50     205  46511 100
#> 13   uuid_13        50     295  25695 100
#> 14   uuid_14        55     377  21033 100
#> 15   uuid_15        46     328  36346 100
#> 16   uuid_16        52     256  22591 100
#> 17   uuid_17        47     277  42873 100
#> 18   uuid_18        54     300  31962 100
#> 19   uuid_19        49     242  27262 100
#> 20   uuid_20        51     351  32126 100
#> 21   uuid_21        49     411  28611 100
#> 22   uuid_22        47     363  26580 100
#> 23   uuid_23        54     296  38400 100
#> 24   uuid_24        50     351  29481 100
#> 25   uuid_25        46     261  43541 100
#> 26   uuid_26        47     294  36804 100
#> 27   uuid_27        45     418  29664 100
#> 28   uuid_28        52     459  43087 100
#> 29   uuid_29        54     451  27276 100
#> 30   uuid_30        46     381  25259 100
#> 31   uuid_31        55     383  48874 100
#> 32   uuid_32        53     370  49925 100
#> 33   uuid_33        47     275  34202 100
#> 34   uuid_34        52     491  41518 100
#> 35   uuid_35        54     365  41060 100
#> 36   uuid_36        48     476  34879 100
#> 37   uuid_37        50     257  28135 100
#> 38   uuid_38        51     351  25491 100
#> 39   uuid_39        47     211  32300 100
#> 40   uuid_40        53     241  33976 100
#> 41   uuid_41        54     306  34077 100
#> 42   uuid_42        49     436  41865 100
#> 43   uuid_43        49     349  34765 100
#> 44   uuid_44        47     280  26319 100
#> 45   uuid_45        52     500  26062 100
#> 46   uuid_46        55     402  29635 100
#> 47   uuid_47        50     340  43710 100
#> 48   uuid_48        55     389  30419 100
#> 49   uuid_49        51     475  46045 100
#> 50   uuid_50        53     471  29809 100
#> 51   uuid_51        45     371  25115 100
#> 52   uuid_52        52     355  45364 100
#> 53   uuid_53        49     486  29875 100
#> 54   uuid_54        48     298  28021 100
#> 55   uuid_55        45     201  41411 100
#> 56   uuid_56        47     287  25261 100
#> 57   uuid_57        49     469  30788 100
#> 58   uuid_58        54     339  37268 100
#> 59   uuid_59        46     277  30740 100
#> 60   uuid_60        49     286  30090 100
#> 61   uuid_61        46     299  44163 100
#> 62   uuid_62        54     328  45795 100
#> 63   uuid_63        49     206  42920 100
#> 64   uuid_64        48     434  41448 100
#> 65   uuid_65        53     392  37149 100
#> 66   uuid_66        47     455  20678 100
#> 67   uuid_67        55     262  24522 100
#> 68   uuid_68        48     490  43358 100
#> 69   uuid_69        48     497  37161 100
#> 70   uuid_70        54     257  47210 100
#> 71   uuid_71        49     209  34349 100
#> 72   uuid_72        48     329  42809 100
#> 73   uuid_73        49     346  44910 100
#> 74   uuid_74        54     326  40954 100
#> 75   uuid_75        46     308  32494 100
#> 76   uuid_76        49     416  42076 100
#> 77   uuid_77        51     223  44918 100
#> 78   uuid_78        51     424  40270 100
#> 79   uuid_79        53     383  37544 100
#> 80   uuid_80        46     336  39632 100
#> 81   uuid_81        52     481  23739 100
#> 82   uuid_82        54     345  33995 100
#> 83   uuid_83        51     269  35440 100
#> 84   uuid_84        47     315  30297 100
#> 85   uuid_85        51     418  45237 100
#> 86   uuid_86        46     201  40747 100
#> 87   uuid_87        47     378  23659 100
#> 88   uuid_88        51     248  30678 100
#> 89   uuid_89        52     228  47808 100
#> 90   uuid_90        46     346  47239 100
#> 91   uuid_91         0     276  45327 100
#> 92   uuid_92         0     266  33139 100
#> 93   uuid_93         1     412  27838 100
#> 94   uuid_94         0     275  29074 100
#> 95   uuid_95         1     430  20957 100
#> 96   uuid_96       100     600  24367 100
#> 97   uuid_97        99     100  31469 100
#> 98   uuid_98       100      80  21005 100
#> 99   uuid_99        99    1020  49906 100
#> 100 uuid_100       100    1050  34729  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         0
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (normal distribution) income           60
#> 16 uuid_2   outlier (normal distribution) income            0
#> 17 uuid_3   outlier (normal distribution) income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>