check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        45     343     60 100
#> 2     uuid_2        53     365      0 100
#> 3     uuid_3        54     356     80 100
#> 4     uuid_4        49     488   1020 100
#> 5     uuid_5        47     409   1050 100
#> 6     uuid_6        48     445  47574 100
#> 7     uuid_7        46     286  22225 100
#> 8     uuid_8        53     452  22487 100
#> 9     uuid_9        46     404  43675 100
#> 10   uuid_10        49     203  28466 100
#> 11   uuid_11        54     367  44761 100
#> 12   uuid_12        46     400  49271 100
#> 13   uuid_13        52     247  21869 100
#> 14   uuid_14        50     228  30301 100
#> 15   uuid_15        49     459  22491 100
#> 16   uuid_16        50     398  43773 100
#> 17   uuid_17        52     461  39838 100
#> 18   uuid_18        46     239  33719 100
#> 19   uuid_19        54     401  30424 100
#> 20   uuid_20        52     466  33508 100
#> 21   uuid_21        48     268  48983 100
#> 22   uuid_22        55     294  20220 100
#> 23   uuid_23        54     323  36615 100
#> 24   uuid_24        53     467  21811 100
#> 25   uuid_25        46     436  48374 100
#> 26   uuid_26        46     232  44070 100
#> 27   uuid_27        45     257  42127 100
#> 28   uuid_28        48     256  37415 100
#> 29   uuid_29        50     287  48199 100
#> 30   uuid_30        51     375  43427 100
#> 31   uuid_31        49     386  49806 100
#> 32   uuid_32        45     411  21150 100
#> 33   uuid_33        50     213  38288 100
#> 34   uuid_34        53     226  38104 100
#> 35   uuid_35        50     289  48451 100
#> 36   uuid_36        55     467  39705 100
#> 37   uuid_37        45     220  27195 100
#> 38   uuid_38        49     329  23408 100
#> 39   uuid_39        48     247  38996 100
#> 40   uuid_40        55     474  22852 100
#> 41   uuid_41        54     358  43936 100
#> 42   uuid_42        50     459  36487 100
#> 43   uuid_43        47     203  47314 100
#> 44   uuid_44        47     386  42407 100
#> 45   uuid_45        54     466  30559 100
#> 46   uuid_46        54     487  44092 100
#> 47   uuid_47        47     388  39169 100
#> 48   uuid_48        51     362  41605 100
#> 49   uuid_49        53     394  30023 100
#> 50   uuid_50        46     405  30285 100
#> 51   uuid_51        46     409  35760 100
#> 52   uuid_52        47     452  35723 100
#> 53   uuid_53        53     258  45783 100
#> 54   uuid_54        46     208  27923 100
#> 55   uuid_55        47     304  24730 100
#> 56   uuid_56        52     345  39260 100
#> 57   uuid_57        52     370  27171 100
#> 58   uuid_58        54     487  33224 100
#> 59   uuid_59        50     362  48509 100
#> 60   uuid_60        50     438  48290 100
#> 61   uuid_61        52     374  29843 100
#> 62   uuid_62        52     225  36410 100
#> 63   uuid_63        45     235  26926 100
#> 64   uuid_64        52     341  36823 100
#> 65   uuid_65        49     227  33195 100
#> 66   uuid_66        51     308  46218 100
#> 67   uuid_67        51     304  48168 100
#> 68   uuid_68        47     265  29517 100
#> 69   uuid_69        50     431  35958 100
#> 70   uuid_70        50     223  25567 100
#> 71   uuid_71        51     476  21170 100
#> 72   uuid_72        52     341  25144 100
#> 73   uuid_73        55     443  24966 100
#> 74   uuid_74        45     395  42058 100
#> 75   uuid_75        49     317  43573 100
#> 76   uuid_76        47     375  39762 100
#> 77   uuid_77        53     403  47601 100
#> 78   uuid_78        50     291  28589 100
#> 79   uuid_79        47     369  49582 100
#> 80   uuid_80        50     255  24942 100
#> 81   uuid_81        49     450  28308 100
#> 82   uuid_82        53     368  46983 100
#> 83   uuid_83        51     427  20935 100
#> 84   uuid_84        46     364  29585 100
#> 85   uuid_85        54     341  39748 100
#> 86   uuid_86        49     246  32406 100
#> 87   uuid_87        50     420  25241 100
#> 88   uuid_88        45     414  44273 100
#> 89   uuid_89        45     317  46021 100
#> 90   uuid_90        49     230  37198 100
#> 91   uuid_91         1     404  25148 100
#> 92   uuid_92         1     484  22284 100
#> 93   uuid_93         0     386  32250 100
#> 94   uuid_94         1     443  21546 100
#> 95   uuid_95         1     499  27162 100
#> 96   uuid_96       100     600  44677 100
#> 97   uuid_97        99     100  35267 100
#> 98   uuid_98        99      80  26600 100
#> 99   uuid_99        99    1020  35246 100
#> 100 uuid_100        99    1050  20389  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>