check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        51     489     60 100
#> 2     uuid_2        50     494      0 100
#> 3     uuid_3        49     316     80 100
#> 4     uuid_4        49     358   1020 100
#> 5     uuid_5        54     320   1050 100
#> 6     uuid_6        55     235  27967 100
#> 7     uuid_7        49     393  20295 100
#> 8     uuid_8        52     425  48090 100
#> 9     uuid_9        47     220  28543 100
#> 10   uuid_10        45     482  34058 100
#> 11   uuid_11        47     414  35935 100
#> 12   uuid_12        54     261  49799 100
#> 13   uuid_13        50     241  42413 100
#> 14   uuid_14        50     348  20074 100
#> 15   uuid_15        50     474  45972 100
#> 16   uuid_16        51     263  46251 100
#> 17   uuid_17        53     226  41619 100
#> 18   uuid_18        47     309  34842 100
#> 19   uuid_19        55     500  32057 100
#> 20   uuid_20        45     416  21397 100
#> 21   uuid_21        49     471  41086 100
#> 22   uuid_22        45     397  26331 100
#> 23   uuid_23        46     253  44422 100
#> 24   uuid_24        53     449  26852 100
#> 25   uuid_25        46     373  34003 100
#> 26   uuid_26        50     232  49153 100
#> 27   uuid_27        55     303  35753 100
#> 28   uuid_28        53     251  49637 100
#> 29   uuid_29        54     443  25034 100
#> 30   uuid_30        54     268  28817 100
#> 31   uuid_31        53     268  29942 100
#> 32   uuid_32        55     287  39065 100
#> 33   uuid_33        54     387  28043 100
#> 34   uuid_34        46     393  41329 100
#> 35   uuid_35        55     483  39011 100
#> 36   uuid_36        50     364  30259 100
#> 37   uuid_37        45     441  29203 100
#> 38   uuid_38        49     341  41531 100
#> 39   uuid_39        48     253  35651 100
#> 40   uuid_40        53     377  44524 100
#> 41   uuid_41        47     220  31226 100
#> 42   uuid_42        48     234  27330 100
#> 43   uuid_43        51     388  29058 100
#> 44   uuid_44        49     239  39431 100
#> 45   uuid_45        49     344  46050 100
#> 46   uuid_46        53     257  25153 100
#> 47   uuid_47        46     268  43005 100
#> 48   uuid_48        55     444  41706 100
#> 49   uuid_49        50     429  22959 100
#> 50   uuid_50        55     454  23200 100
#> 51   uuid_51        46     335  39781 100
#> 52   uuid_52        51     224  42037 100
#> 53   uuid_53        47     322  26169 100
#> 54   uuid_54        47     276  44447 100
#> 55   uuid_55        52     231  43591 100
#> 56   uuid_56        55     206  30954 100
#> 57   uuid_57        53     271  30420 100
#> 58   uuid_58        54     492  38374 100
#> 59   uuid_59        51     460  39180 100
#> 60   uuid_60        47     253  39261 100
#> 61   uuid_61        53     257  31672 100
#> 62   uuid_62        52     239  21697 100
#> 63   uuid_63        52     449  46111 100
#> 64   uuid_64        46     417  45556 100
#> 65   uuid_65        52     256  38966 100
#> 66   uuid_66        51     427  47324 100
#> 67   uuid_67        48     479  21086 100
#> 68   uuid_68        50     451  20027 100
#> 69   uuid_69        53     459  31591 100
#> 70   uuid_70        55     257  44201 100
#> 71   uuid_71        49     357  30687 100
#> 72   uuid_72        54     310  33905 100
#> 73   uuid_73        49     339  35042 100
#> 74   uuid_74        51     289  46441 100
#> 75   uuid_75        52     395  30769 100
#> 76   uuid_76        46     475  43883 100
#> 77   uuid_77        46     444  23951 100
#> 78   uuid_78        47     382  20191 100
#> 79   uuid_79        55     345  38990 100
#> 80   uuid_80        47     302  45915 100
#> 81   uuid_81        49     455  44917 100
#> 82   uuid_82        49     321  40570 100
#> 83   uuid_83        48     261  22672 100
#> 84   uuid_84        54     453  24832 100
#> 85   uuid_85        50     422  36552 100
#> 86   uuid_86        55     440  33289 100
#> 87   uuid_87        53     334  45575 100
#> 88   uuid_88        53     366  23055 100
#> 89   uuid_89        50     294  40959 100
#> 90   uuid_90        55     260  21123 100
#> 91   uuid_91         0     211  21641 100
#> 92   uuid_92         1     329  40164 100
#> 93   uuid_93         1     495  27406 100
#> 94   uuid_94         0     468  40886 100
#> 95   uuid_95         1     484  46595 100
#> 96   uuid_96        99     600  30935 100
#> 97   uuid_97        99     100  32662 100
#> 98   uuid_98        99      80  37198 100
#> 99   uuid_99        99    1020  26012 100
#> 100 uuid_100        99    1050  34130  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>