check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        53     365     60 100
#> 2     uuid_2        46     375      0 100
#> 3     uuid_3        53     261     80 100
#> 4     uuid_4        51     366   1020 100
#> 5     uuid_5        51     449   1050 100
#> 6     uuid_6        51     427  22335 100
#> 7     uuid_7        52     205  26746 100
#> 8     uuid_8        55     289  35868 100
#> 9     uuid_9        50     462  24884 100
#> 10   uuid_10        54     276  40257 100
#> 11   uuid_11        45     482  45384 100
#> 12   uuid_12        50     204  38373 100
#> 13   uuid_13        50     360  41022 100
#> 14   uuid_14        51     293  37522 100
#> 15   uuid_15        51     327  43287 100
#> 16   uuid_16        51     500  32522 100
#> 17   uuid_17        54     279  29750 100
#> 18   uuid_18        54     363  38465 100
#> 19   uuid_19        50     206  35311 100
#> 20   uuid_20        46     380  33024 100
#> 21   uuid_21        49     473  32029 100
#> 22   uuid_22        49     266  26197 100
#> 23   uuid_23        54     309  21443 100
#> 24   uuid_24        49     422  38740 100
#> 25   uuid_25        47     430  20259 100
#> 26   uuid_26        47     442  34328 100
#> 27   uuid_27        54     342  48957 100
#> 28   uuid_28        51     292  46535 100
#> 29   uuid_29        48     352  40799 100
#> 30   uuid_30        52     480  42999 100
#> 31   uuid_31        48     404  45085 100
#> 32   uuid_32        46     208  49441 100
#> 33   uuid_33        46     405  25174 100
#> 34   uuid_34        48     360  44355 100
#> 35   uuid_35        50     306  26494 100
#> 36   uuid_36        51     436  46878 100
#> 37   uuid_37        52     209  31117 100
#> 38   uuid_38        46     475  42800 100
#> 39   uuid_39        48     292  34998 100
#> 40   uuid_40        49     317  41827 100
#> 41   uuid_41        52     252  47452 100
#> 42   uuid_42        54     239  46106 100
#> 43   uuid_43        55     391  30616 100
#> 44   uuid_44        48     278  22418 100
#> 45   uuid_45        50     205  40476 100
#> 46   uuid_46        50     401  21217 100
#> 47   uuid_47        54     265  40232 100
#> 48   uuid_48        51     428  33030 100
#> 49   uuid_49        52     463  43705 100
#> 50   uuid_50        52     499  40553 100
#> 51   uuid_51        51     286  26131 100
#> 52   uuid_52        54     264  49136 100
#> 53   uuid_53        55     427  27570 100
#> 54   uuid_54        46     483  44184 100
#> 55   uuid_55        49     285  40769 100
#> 56   uuid_56        54     370  46671 100
#> 57   uuid_57        49     394  40807 100
#> 58   uuid_58        48     204  27679 100
#> 59   uuid_59        51     338  34429 100
#> 60   uuid_60        52     347  41979 100
#> 61   uuid_61        49     218  41555 100
#> 62   uuid_62        51     499  20372 100
#> 63   uuid_63        51     489  20985 100
#> 64   uuid_64        47     269  20255 100
#> 65   uuid_65        52     332  35570 100
#> 66   uuid_66        49     366  39694 100
#> 67   uuid_67        49     415  29825 100
#> 68   uuid_68        55     293  35340 100
#> 69   uuid_69        51     291  39635 100
#> 70   uuid_70        52     367  36785 100
#> 71   uuid_71        53     257  25027 100
#> 72   uuid_72        48     250  43677 100
#> 73   uuid_73        47     340  30912 100
#> 74   uuid_74        48     355  21964 100
#> 75   uuid_75        48     307  29345 100
#> 76   uuid_76        49     343  36886 100
#> 77   uuid_77        46     396  21697 100
#> 78   uuid_78        48     200  35016 100
#> 79   uuid_79        49     461  30472 100
#> 80   uuid_80        48     460  43846 100
#> 81   uuid_81        52     368  30965 100
#> 82   uuid_82        49     453  49002 100
#> 83   uuid_83        54     273  44266 100
#> 84   uuid_84        49     272  42712 100
#> 85   uuid_85        53     232  45731 100
#> 86   uuid_86        54     382  28098 100
#> 87   uuid_87        54     243  32148 100
#> 88   uuid_88        49     424  27549 100
#> 89   uuid_89        54     257  23482 100
#> 90   uuid_90        53     247  37210 100
#> 91   uuid_91         1     452  23704 100
#> 92   uuid_92         0     233  34577 100
#> 93   uuid_93         0     302  34033 100
#> 94   uuid_94         0     342  31507 100
#> 95   uuid_95         0     294  24267 100
#> 96   uuid_96        99     600  27795 100
#> 97   uuid_97        99     100  40674 100
#> 98   uuid_98       100      80  22485 100
#> 99   uuid_99       100    1020  36488 100
#> 100 uuid_100       100    1050  41195  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         0
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>