check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        54     478     60 100
#> 2     uuid_2        54     399      0 100
#> 3     uuid_3        45     454     80 100
#> 4     uuid_4        45     304   1020 100
#> 5     uuid_5        50     463   1050 100
#> 6     uuid_6        49     411  35918 100
#> 7     uuid_7        45     445  30517 100
#> 8     uuid_8        55     220  38966 100
#> 9     uuid_9        46     319  27888 100
#> 10   uuid_10        47     444  36858 100
#> 11   uuid_11        53     286  33917 100
#> 12   uuid_12        55     217  49098 100
#> 13   uuid_13        47     311  44282 100
#> 14   uuid_14        52     409  24080 100
#> 15   uuid_15        54     448  20171 100
#> 16   uuid_16        48     365  35888 100
#> 17   uuid_17        54     246  21846 100
#> 18   uuid_18        54     312  49817 100
#> 19   uuid_19        54     341  36860 100
#> 20   uuid_20        51     258  48360 100
#> 21   uuid_21        49     441  41790 100
#> 22   uuid_22        51     299  28787 100
#> 23   uuid_23        50     473  35464 100
#> 24   uuid_24        51     403  43386 100
#> 25   uuid_25        48     243  38652 100
#> 26   uuid_26        48     280  38676 100
#> 27   uuid_27        51     445  30413 100
#> 28   uuid_28        45     297  29418 100
#> 29   uuid_29        49     247  44131 100
#> 30   uuid_30        46     440  47768 100
#> 31   uuid_31        45     289  36560 100
#> 32   uuid_32        54     464  33952 100
#> 33   uuid_33        49     378  48261 100
#> 34   uuid_34        47     225  39902 100
#> 35   uuid_35        48     283  37019 100
#> 36   uuid_36        46     210  20693 100
#> 37   uuid_37        48     259  36156 100
#> 38   uuid_38        47     269  24610 100
#> 39   uuid_39        47     279  30744 100
#> 40   uuid_40        46     447  47804 100
#> 41   uuid_41        46     257  42015 100
#> 42   uuid_42        48     263  45655 100
#> 43   uuid_43        51     295  37381 100
#> 44   uuid_44        53     310  27669 100
#> 45   uuid_45        47     252  22567 100
#> 46   uuid_46        49     456  33306 100
#> 47   uuid_47        46     215  34068 100
#> 48   uuid_48        47     279  38493 100
#> 49   uuid_49        50     293  46928 100
#> 50   uuid_50        46     351  47699 100
#> 51   uuid_51        49     248  28978 100
#> 52   uuid_52        54     470  22262 100
#> 53   uuid_53        55     252  22736 100
#> 54   uuid_54        51     377  49289 100
#> 55   uuid_55        47     391  23402 100
#> 56   uuid_56        49     330  28870 100
#> 57   uuid_57        50     234  26431 100
#> 58   uuid_58        53     462  22371 100
#> 59   uuid_59        54     461  42229 100
#> 60   uuid_60        49     383  32941 100
#> 61   uuid_61        47     460  42199 100
#> 62   uuid_62        52     425  39872 100
#> 63   uuid_63        48     474  45253 100
#> 64   uuid_64        48     494  36762 100
#> 65   uuid_65        47     314  28385 100
#> 66   uuid_66        52     354  36922 100
#> 67   uuid_67        46     473  49457 100
#> 68   uuid_68        51     305  40625 100
#> 69   uuid_69        49     218  24939 100
#> 70   uuid_70        54     428  25014 100
#> 71   uuid_71        49     308  28936 100
#> 72   uuid_72        46     372  23131 100
#> 73   uuid_73        52     208  23103 100
#> 74   uuid_74        51     459  22162 100
#> 75   uuid_75        50     321  47413 100
#> 76   uuid_76        50     476  30605 100
#> 77   uuid_77        52     401  37827 100
#> 78   uuid_78        45     305  45585 100
#> 79   uuid_79        47     472  25203 100
#> 80   uuid_80        47     490  42797 100
#> 81   uuid_81        54     317  46466 100
#> 82   uuid_82        54     415  42313 100
#> 83   uuid_83        54     217  24980 100
#> 84   uuid_84        53     393  36082 100
#> 85   uuid_85        51     255  46973 100
#> 86   uuid_86        53     330  25085 100
#> 87   uuid_87        51     379  33525 100
#> 88   uuid_88        51     297  37264 100
#> 89   uuid_89        50     473  37163 100
#> 90   uuid_90        49     453  33734 100
#> 91   uuid_91         1     325  41961 100
#> 92   uuid_92         0     383  29228 100
#> 93   uuid_93         0     373  30237 100
#> 94   uuid_94         1     287  26628 100
#> 95   uuid_95         0     297  35985 100
#> 96   uuid_96       100     600  38058 100
#> 97   uuid_97        99     100  44811 100
#> 98   uuid_98       100      80  33200 100
#> 99   uuid_99        99    1020  40555 100
#> 100 uuid_100       100    1050  23264  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         0
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>