check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        54     463     60 100
#> 2     uuid_2        54     264      0 100
#> 3     uuid_3        51     262     80 100
#> 4     uuid_4        52     471   1020 100
#> 5     uuid_5        54     471   1050 100
#> 6     uuid_6        55     434  33100 100
#> 7     uuid_7        52     457  22300 100
#> 8     uuid_8        51     294  24696 100
#> 9     uuid_9        51     311  32303 100
#> 10   uuid_10        51     450  43994 100
#> 11   uuid_11        49     243  28872 100
#> 12   uuid_12        47     342  24329 100
#> 13   uuid_13        47     485  26195 100
#> 14   uuid_14        46     420  32868 100
#> 15   uuid_15        54     277  32871 100
#> 16   uuid_16        48     497  33883 100
#> 17   uuid_17        48     472  28990 100
#> 18   uuid_18        54     389  33931 100
#> 19   uuid_19        51     458  45099 100
#> 20   uuid_20        51     498  31423 100
#> 21   uuid_21        51     270  28119 100
#> 22   uuid_22        50     368  46229 100
#> 23   uuid_23        54     320  49037 100
#> 24   uuid_24        53     344  33043 100
#> 25   uuid_25        53     396  24653 100
#> 26   uuid_26        46     210  45798 100
#> 27   uuid_27        48     457  34087 100
#> 28   uuid_28        50     279  21781 100
#> 29   uuid_29        49     366  40942 100
#> 30   uuid_30        52     399  49730 100
#> 31   uuid_31        49     372  20976 100
#> 32   uuid_32        53     459  29500 100
#> 33   uuid_33        53     309  46367 100
#> 34   uuid_34        53     216  25676 100
#> 35   uuid_35        50     444  38545 100
#> 36   uuid_36        53     357  25690 100
#> 37   uuid_37        47     201  31332 100
#> 38   uuid_38        45     300  35764 100
#> 39   uuid_39        49     293  40467 100
#> 40   uuid_40        50     322  34130 100
#> 41   uuid_41        54     332  21321 100
#> 42   uuid_42        45     314  21837 100
#> 43   uuid_43        49     235  29844 100
#> 44   uuid_44        47     238  34400 100
#> 45   uuid_45        52     432  36170 100
#> 46   uuid_46        45     386  20407 100
#> 47   uuid_47        45     379  35186 100
#> 48   uuid_48        54     321  35564 100
#> 49   uuid_49        53     416  21908 100
#> 50   uuid_50        50     456  42316 100
#> 51   uuid_51        47     356  43665 100
#> 52   uuid_52        47     247  44819 100
#> 53   uuid_53        49     341  29445 100
#> 54   uuid_54        46     363  37532 100
#> 55   uuid_55        47     358  42150 100
#> 56   uuid_56        51     489  27589 100
#> 57   uuid_57        53     382  39728 100
#> 58   uuid_58        55     397  39149 100
#> 59   uuid_59        46     365  31852 100
#> 60   uuid_60        54     296  21142 100
#> 61   uuid_61        46     249  40349 100
#> 62   uuid_62        50     375  45697 100
#> 63   uuid_63        49     301  48805 100
#> 64   uuid_64        51     393  29402 100
#> 65   uuid_65        45     426  31049 100
#> 66   uuid_66        53     220  43194 100
#> 67   uuid_67        47     328  24075 100
#> 68   uuid_68        53     233  47773 100
#> 69   uuid_69        52     395  24180 100
#> 70   uuid_70        46     457  29479 100
#> 71   uuid_71        50     423  35986 100
#> 72   uuid_72        53     490  49550 100
#> 73   uuid_73        51     214  41236 100
#> 74   uuid_74        48     287  29248 100
#> 75   uuid_75        54     350  47886 100
#> 76   uuid_76        51     418  29396 100
#> 77   uuid_77        50     333  48355 100
#> 78   uuid_78        52     459  45657 100
#> 79   uuid_79        48     489  41236 100
#> 80   uuid_80        48     327  27791 100
#> 81   uuid_81        49     456  34899 100
#> 82   uuid_82        45     470  25714 100
#> 83   uuid_83        49     309  26018 100
#> 84   uuid_84        51     394  43364 100
#> 85   uuid_85        48     405  46118 100
#> 86   uuid_86        45     452  46467 100
#> 87   uuid_87        46     465  42435 100
#> 88   uuid_88        49     399  23930 100
#> 89   uuid_89        54     327  22010 100
#> 90   uuid_90        48     480  26096 100
#> 91   uuid_91         1     229  21591 100
#> 92   uuid_92         1     448  25999 100
#> 93   uuid_93         1     235  32049 100
#> 94   uuid_94         0     265  32230 100
#> 95   uuid_95         1     407  39034 100
#> 96   uuid_96       100     600  49695 100
#> 97   uuid_97        99     100  21132 100
#> 98   uuid_98        99      80  43634 100
#> 99   uuid_99       100    1020  24918 100
#> 100 uuid_100       100    1050  38445  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>