check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        50     440     60 100
#> 2     uuid_2        52     221      0 100
#> 3     uuid_3        53     269     80 100
#> 4     uuid_4        50     346   1020 100
#> 5     uuid_5        52     387   1050 100
#> 6     uuid_6        47     312  30298 100
#> 7     uuid_7        53     363  23246 100
#> 8     uuid_8        47     330  32453 100
#> 9     uuid_9        47     481  40839 100
#> 10   uuid_10        51     417  27097 100
#> 11   uuid_11        48     255  42437 100
#> 12   uuid_12        51     362  36651 100
#> 13   uuid_13        48     230  22249 100
#> 14   uuid_14        53     283  20745 100
#> 15   uuid_15        53     226  20256 100
#> 16   uuid_16        53     211  23176 100
#> 17   uuid_17        49     352  24518 100
#> 18   uuid_18        55     308  32192 100
#> 19   uuid_19        52     493  30916 100
#> 20   uuid_20        46     368  49799 100
#> 21   uuid_21        53     207  35872 100
#> 22   uuid_22        52     250  26839 100
#> 23   uuid_23        53     259  27634 100
#> 24   uuid_24        51     335  30977 100
#> 25   uuid_25        49     422  33975 100
#> 26   uuid_26        54     333  24316 100
#> 27   uuid_27        50     349  31141 100
#> 28   uuid_28        55     495  38608 100
#> 29   uuid_29        55     406  34296 100
#> 30   uuid_30        53     212  47450 100
#> 31   uuid_31        55     492  48392 100
#> 32   uuid_32        51     266  24871 100
#> 33   uuid_33        48     432  20488 100
#> 34   uuid_34        45     488  21328 100
#> 35   uuid_35        49     357  44907 100
#> 36   uuid_36        50     413  46570 100
#> 37   uuid_37        49     396  43089 100
#> 38   uuid_38        45     499  36697 100
#> 39   uuid_39        52     358  46605 100
#> 40   uuid_40        50     453  42213 100
#> 41   uuid_41        47     298  31572 100
#> 42   uuid_42        47     289  43754 100
#> 43   uuid_43        53     393  24957 100
#> 44   uuid_44        47     356  29900 100
#> 45   uuid_45        46     264  30764 100
#> 46   uuid_46        54     462  40986 100
#> 47   uuid_47        45     380  40340 100
#> 48   uuid_48        46     376  34077 100
#> 49   uuid_49        50     387  37023 100
#> 50   uuid_50        51     230  25890 100
#> 51   uuid_51        46     390  46268 100
#> 52   uuid_52        51     486  33780 100
#> 53   uuid_53        47     489  22183 100
#> 54   uuid_54        45     255  36079 100
#> 55   uuid_55        52     280  49219 100
#> 56   uuid_56        52     480  26054 100
#> 57   uuid_57        52     439  31735 100
#> 58   uuid_58        50     470  21117 100
#> 59   uuid_59        52     357  34175 100
#> 60   uuid_60        46     383  39857 100
#> 61   uuid_61        52     420  27941 100
#> 62   uuid_62        47     339  26262 100
#> 63   uuid_63        51     497  40817 100
#> 64   uuid_64        55     470  37825 100
#> 65   uuid_65        54     356  30020 100
#> 66   uuid_66        55     362  36047 100
#> 67   uuid_67        51     451  26688 100
#> 68   uuid_68        45     486  33281 100
#> 69   uuid_69        52     463  30310 100
#> 70   uuid_70        53     362  46962 100
#> 71   uuid_71        47     455  43590 100
#> 72   uuid_72        49     488  45654 100
#> 73   uuid_73        50     311  32754 100
#> 74   uuid_74        52     421  48388 100
#> 75   uuid_75        46     494  35246 100
#> 76   uuid_76        47     216  25989 100
#> 77   uuid_77        48     477  46209 100
#> 78   uuid_78        50     341  31103 100
#> 79   uuid_79        54     413  37845 100
#> 80   uuid_80        47     290  25455 100
#> 81   uuid_81        51     311  30611 100
#> 82   uuid_82        52     463  24639 100
#> 83   uuid_83        52     223  34113 100
#> 84   uuid_84        50     449  22847 100
#> 85   uuid_85        50     220  23091 100
#> 86   uuid_86        45     452  29406 100
#> 87   uuid_87        52     349  40023 100
#> 88   uuid_88        46     327  21435 100
#> 89   uuid_89        51     352  30646 100
#> 90   uuid_90        54     347  26968 100
#> 91   uuid_91         1     263  34486 100
#> 92   uuid_92         1     334  44862 100
#> 93   uuid_93         0     461  39684 100
#> 94   uuid_94         1     406  36042 100
#> 95   uuid_95         0     479  41540 100
#> 96   uuid_96        99     600  49575 100
#> 97   uuid_97        99     100  45307 100
#> 98   uuid_98       100      80  40337 100
#> 99   uuid_99       100    1020  32823 100
#> 100 uuid_100       100    1050  31020  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>