check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        47     237     60 100
#> 2     uuid_2        51     282      0 100
#> 3     uuid_3        53     293     80 100
#> 4     uuid_4        45     305   1020 100
#> 5     uuid_5        51     228   1050 100
#> 6     uuid_6        50     329  44636 100
#> 7     uuid_7        50     233  25589 100
#> 8     uuid_8        47     409  37740 100
#> 9     uuid_9        51     249  45212 100
#> 10   uuid_10        50     375  26539 100
#> 11   uuid_11        47     311  48985 100
#> 12   uuid_12        51     255  29598 100
#> 13   uuid_13        46     486  32059 100
#> 14   uuid_14        45     264  44497 100
#> 15   uuid_15        52     237  25894 100
#> 16   uuid_16        48     468  31932 100
#> 17   uuid_17        46     288  21575 100
#> 18   uuid_18        47     436  25379 100
#> 19   uuid_19        53     498  38186 100
#> 20   uuid_20        53     316  31585 100
#> 21   uuid_21        50     246  38534 100
#> 22   uuid_22        52     272  29443 100
#> 23   uuid_23        54     361  41374 100
#> 24   uuid_24        48     224  34690 100
#> 25   uuid_25        47     268  32164 100
#> 26   uuid_26        52     204  34630 100
#> 27   uuid_27        51     277  20718 100
#> 28   uuid_28        46     325  41941 100
#> 29   uuid_29        45     418  46395 100
#> 30   uuid_30        46     272  36694 100
#> 31   uuid_31        49     431  24636 100
#> 32   uuid_32        54     393  47834 100
#> 33   uuid_33        50     276  30824 100
#> 34   uuid_34        52     285  32501 100
#> 35   uuid_35        53     287  27054 100
#> 36   uuid_36        54     293  20768 100
#> 37   uuid_37        48     254  23291 100
#> 38   uuid_38        48     208  36116 100
#> 39   uuid_39        52     338  36129 100
#> 40   uuid_40        53     352  44847 100
#> 41   uuid_41        47     471  26376 100
#> 42   uuid_42        50     247  46139 100
#> 43   uuid_43        50     406  24213 100
#> 44   uuid_44        50     316  43131 100
#> 45   uuid_45        55     367  49523 100
#> 46   uuid_46        54     484  34828 100
#> 47   uuid_47        54     229  42745 100
#> 48   uuid_48        55     238  23765 100
#> 49   uuid_49        53     238  22544 100
#> 50   uuid_50        50     339  24399 100
#> 51   uuid_51        52     446  39845 100
#> 52   uuid_52        53     322  30484 100
#> 53   uuid_53        48     381  35395 100
#> 54   uuid_54        52     488  40952 100
#> 55   uuid_55        45     474  35691 100
#> 56   uuid_56        54     220  49347 100
#> 57   uuid_57        52     209  35788 100
#> 58   uuid_58        53     412  22615 100
#> 59   uuid_59        53     323  41076 100
#> 60   uuid_60        52     407  40321 100
#> 61   uuid_61        49     387  23349 100
#> 62   uuid_62        55     284  35684 100
#> 63   uuid_63        47     493  24231 100
#> 64   uuid_64        47     478  29055 100
#> 65   uuid_65        53     273  34159 100
#> 66   uuid_66        53     395  32792 100
#> 67   uuid_67        46     411  43957 100
#> 68   uuid_68        47     288  43627 100
#> 69   uuid_69        54     242  33271 100
#> 70   uuid_70        48     474  39381 100
#> 71   uuid_71        50     206  45932 100
#> 72   uuid_72        48     253  41153 100
#> 73   uuid_73        52     447  20234 100
#> 74   uuid_74        47     204  45876 100
#> 75   uuid_75        54     331  43036 100
#> 76   uuid_76        53     393  20125 100
#> 77   uuid_77        53     407  27592 100
#> 78   uuid_78        47     401  40323 100
#> 79   uuid_79        52     431  22320 100
#> 80   uuid_80        45     446  40181 100
#> 81   uuid_81        54     442  34344 100
#> 82   uuid_82        53     267  42181 100
#> 83   uuid_83        52     441  44359 100
#> 84   uuid_84        50     381  48596 100
#> 85   uuid_85        46     486  20897 100
#> 86   uuid_86        52     261  39376 100
#> 87   uuid_87        45     261  26043 100
#> 88   uuid_88        47     295  22439 100
#> 89   uuid_89        52     367  37420 100
#> 90   uuid_90        47     426  48893 100
#> 91   uuid_91         1     323  22536 100
#> 92   uuid_92         1     283  45150 100
#> 93   uuid_93         0     428  21472 100
#> 94   uuid_94         1     212  43647 100
#> 95   uuid_95         0     486  29563 100
#> 96   uuid_96       100     600  43423 100
#> 97   uuid_97       100     100  24680 100
#> 98   uuid_98       100      80  44017 100
#> 99   uuid_99        99    1020  29805 100
#> 100 uuid_100        99    1050  30623  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value       100
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>