check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        52     318     60 100
#> 2     uuid_2        52     456      0 100
#> 3     uuid_3        53     393     80 100
#> 4     uuid_4        53     359   1020 100
#> 5     uuid_5        54     227   1050 100
#> 6     uuid_6        48     252  33028 100
#> 7     uuid_7        53     391  42507 100
#> 8     uuid_8        52     489  22433 100
#> 9     uuid_9        50     472  35803 100
#> 10   uuid_10        51     364  28809 100
#> 11   uuid_11        54     327  37528 100
#> 12   uuid_12        51     258  33341 100
#> 13   uuid_13        54     480  26028 100
#> 14   uuid_14        54     344  24338 100
#> 15   uuid_15        49     398  38754 100
#> 16   uuid_16        47     422  49337 100
#> 17   uuid_17        49     280  27212 100
#> 18   uuid_18        52     400  28020 100
#> 19   uuid_19        50     312  25155 100
#> 20   uuid_20        52     359  23994 100
#> 21   uuid_21        49     483  28737 100
#> 22   uuid_22        46     334  30306 100
#> 23   uuid_23        53     436  42360 100
#> 24   uuid_24        46     436  26838 100
#> 25   uuid_25        54     465  27261 100
#> 26   uuid_26        50     216  46845 100
#> 27   uuid_27        52     404  43763 100
#> 28   uuid_28        54     337  29149 100
#> 29   uuid_29        52     273  33004 100
#> 30   uuid_30        47     310  45368 100
#> 31   uuid_31        50     238  43359 100
#> 32   uuid_32        52     214  28100 100
#> 33   uuid_33        47     389  24517 100
#> 34   uuid_34        49     381  23074 100
#> 35   uuid_35        53     218  25318 100
#> 36   uuid_36        45     412  20241 100
#> 37   uuid_37        50     442  46001 100
#> 38   uuid_38        50     333  45432 100
#> 39   uuid_39        52     425  28656 100
#> 40   uuid_40        55     272  37141 100
#> 41   uuid_41        54     313  39474 100
#> 42   uuid_42        53     231  24678 100
#> 43   uuid_43        46     481  31692 100
#> 44   uuid_44        55     455  37785 100
#> 45   uuid_45        46     465  30081 100
#> 46   uuid_46        50     220  32708 100
#> 47   uuid_47        46     421  24356 100
#> 48   uuid_48        52     352  20139 100
#> 49   uuid_49        49     484  47529 100
#> 50   uuid_50        51     352  49771 100
#> 51   uuid_51        51     241  47148 100
#> 52   uuid_52        53     295  41095 100
#> 53   uuid_53        49     341  24601 100
#> 54   uuid_54        53     406  27602 100
#> 55   uuid_55        47     485  48781 100
#> 56   uuid_56        51     322  25236 100
#> 57   uuid_57        55     407  26975 100
#> 58   uuid_58        51     345  22140 100
#> 59   uuid_59        48     308  43329 100
#> 60   uuid_60        55     329  26033 100
#> 61   uuid_61        46     423  44739 100
#> 62   uuid_62        47     263  33347 100
#> 63   uuid_63        52     299  48517 100
#> 64   uuid_64        45     297  45727 100
#> 65   uuid_65        48     356  48061 100
#> 66   uuid_66        45     215  39981 100
#> 67   uuid_67        55     222  44758 100
#> 68   uuid_68        52     492  49502 100
#> 69   uuid_69        46     277  21967 100
#> 70   uuid_70        52     445  20075 100
#> 71   uuid_71        53     351  36094 100
#> 72   uuid_72        49     480  46735 100
#> 73   uuid_73        47     486  27010 100
#> 74   uuid_74        49     381  21684 100
#> 75   uuid_75        48     367  36335 100
#> 76   uuid_76        45     377  37889 100
#> 77   uuid_77        49     436  29252 100
#> 78   uuid_78        48     483  32162 100
#> 79   uuid_79        46     497  45497 100
#> 80   uuid_80        51     204  47470 100
#> 81   uuid_81        53     317  22209 100
#> 82   uuid_82        52     209  20177 100
#> 83   uuid_83        50     307  44475 100
#> 84   uuid_84        53     278  40245 100
#> 85   uuid_85        54     256  30806 100
#> 86   uuid_86        47     469  42247 100
#> 87   uuid_87        46     347  42798 100
#> 88   uuid_88        50     363  48152 100
#> 89   uuid_89        45     274  38224 100
#> 90   uuid_90        54     344  25327 100
#> 91   uuid_91         1     314  47781 100
#> 92   uuid_92         1     464  22263 100
#> 93   uuid_93         0     489  47159 100
#> 94   uuid_94         1     430  34621 100
#> 95   uuid_95         1     301  43284 100
#> 96   uuid_96        99     600  47248 100
#> 97   uuid_97       100     100  21248 100
#> 98   uuid_98       100      80  48102 100
#> 99   uuid_99        99    1020  45661 100
#> 100 uuid_100        99    1050  33920  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value       100
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value        99
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>