check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        48     355     60 100
#> 2     uuid_2        54     493      0 100
#> 3     uuid_3        47     442     80 100
#> 4     uuid_4        54     435   1020 100
#> 5     uuid_5        46     428   1050 100
#> 6     uuid_6        49     229  39903 100
#> 7     uuid_7        55     457  20816 100
#> 8     uuid_8        51     497  27974 100
#> 9     uuid_9        46     265  23708 100
#> 10   uuid_10        54     412  45425 100
#> 11   uuid_11        52     394  39782 100
#> 12   uuid_12        47     292  32609 100
#> 13   uuid_13        46     464  38447 100
#> 14   uuid_14        45     228  47622 100
#> 15   uuid_15        47     258  44692 100
#> 16   uuid_16        47     434  36901 100
#> 17   uuid_17        49     344  30072 100
#> 18   uuid_18        50     471  44418 100
#> 19   uuid_19        54     430  24081 100
#> 20   uuid_20        46     402  41834 100
#> 21   uuid_21        48     414  39696 100
#> 22   uuid_22        53     221  46380 100
#> 23   uuid_23        46     403  31693 100
#> 24   uuid_24        45     341  23203 100
#> 25   uuid_25        47     441  41144 100
#> 26   uuid_26        54     347  32521 100
#> 27   uuid_27        55     290  37113 100
#> 28   uuid_28        52     457  21513 100
#> 29   uuid_29        51     492  32926 100
#> 30   uuid_30        48     244  37040 100
#> 31   uuid_31        52     487  46794 100
#> 32   uuid_32        48     479  38526 100
#> 33   uuid_33        54     282  22194 100
#> 34   uuid_34        55     435  48620 100
#> 35   uuid_35        49     286  25694 100
#> 36   uuid_36        51     391  35574 100
#> 37   uuid_37        45     347  27699 100
#> 38   uuid_38        49     215  31220 100
#> 39   uuid_39        54     434  30202 100
#> 40   uuid_40        47     355  37854 100
#> 41   uuid_41        54     403  36392 100
#> 42   uuid_42        54     258  23760 100
#> 43   uuid_43        47     322  39478 100
#> 44   uuid_44        46     426  21666 100
#> 45   uuid_45        48     447  33008 100
#> 46   uuid_46        53     285  20828 100
#> 47   uuid_47        48     214  49158 100
#> 48   uuid_48        46     362  28096 100
#> 49   uuid_49        49     306  22742 100
#> 50   uuid_50        51     413  45751 100
#> 51   uuid_51        48     246  24882 100
#> 52   uuid_52        50     482  33929 100
#> 53   uuid_53        48     363  41633 100
#> 54   uuid_54        53     478  42121 100
#> 55   uuid_55        55     447  48381 100
#> 56   uuid_56        50     238  29112 100
#> 57   uuid_57        51     351  21203 100
#> 58   uuid_58        46     437  40603 100
#> 59   uuid_59        46     374  46771 100
#> 60   uuid_60        50     228  49436 100
#> 61   uuid_61        53     415  46985 100
#> 62   uuid_62        52     476  37974 100
#> 63   uuid_63        52     352  20399 100
#> 64   uuid_64        53     207  30626 100
#> 65   uuid_65        52     318  42052 100
#> 66   uuid_66        53     499  41375 100
#> 67   uuid_67        51     485  47303 100
#> 68   uuid_68        47     311  28101 100
#> 69   uuid_69        45     444  22333 100
#> 70   uuid_70        55     466  31802 100
#> 71   uuid_71        47     375  20610 100
#> 72   uuid_72        46     277  27587 100
#> 73   uuid_73        54     380  22629 100
#> 74   uuid_74        51     246  46616 100
#> 75   uuid_75        50     487  24244 100
#> 76   uuid_76        53     360  26734 100
#> 77   uuid_77        53     473  34788 100
#> 78   uuid_78        49     288  28634 100
#> 79   uuid_79        53     484  43619 100
#> 80   uuid_80        51     474  20670 100
#> 81   uuid_81        51     253  32135 100
#> 82   uuid_82        45     325  36077 100
#> 83   uuid_83        50     406  29690 100
#> 84   uuid_84        46     269  24556 100
#> 85   uuid_85        53     434  42960 100
#> 86   uuid_86        47     409  47098 100
#> 87   uuid_87        47     219  24764 100
#> 88   uuid_88        49     353  34664 100
#> 89   uuid_89        52     494  28271 100
#> 90   uuid_90        51     248  24912 100
#> 91   uuid_91         0     455  48410 100
#> 92   uuid_92         1     237  46058 100
#> 93   uuid_93         1     290  39298 100
#> 94   uuid_94         1     218  33487 100
#> 95   uuid_95         0     365  37128 100
#> 96   uuid_96        99     600  33053 100
#> 97   uuid_97        99     100  43334 100
#> 98   uuid_98        99      80  26431 100
#> 99   uuid_99       100    1020  43083 100
#> 100 uuid_100        99    1050  21201  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>