check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        53     322     60 100
#> 2     uuid_2        46     422      0 100
#> 3     uuid_3        54     252     80 100
#> 4     uuid_4        47     290   1020 100
#> 5     uuid_5        47     350   1050 100
#> 6     uuid_6        45     342  40633 100
#> 7     uuid_7        53     311  25514 100
#> 8     uuid_8        52     256  41755 100
#> 9     uuid_9        46     249  20095 100
#> 10   uuid_10        49     281  26605 100
#> 11   uuid_11        53     204  22711 100
#> 12   uuid_12        48     482  43610 100
#> 13   uuid_13        52     306  45697 100
#> 14   uuid_14        52     478  24589 100
#> 15   uuid_15        53     395  29213 100
#> 16   uuid_16        45     469  31257 100
#> 17   uuid_17        52     251  37435 100
#> 18   uuid_18        45     287  48166 100
#> 19   uuid_19        55     289  31471 100
#> 20   uuid_20        54     348  27943 100
#> 21   uuid_21        48     331  22287 100
#> 22   uuid_22        53     222  35314 100
#> 23   uuid_23        50     481  43893 100
#> 24   uuid_24        47     413  22428 100
#> 25   uuid_25        55     283  28750 100
#> 26   uuid_26        48     359  32342 100
#> 27   uuid_27        54     224  37180 100
#> 28   uuid_28        53     232  37824 100
#> 29   uuid_29        53     362  48043 100
#> 30   uuid_30        48     259  32880 100
#> 31   uuid_31        46     334  35418 100
#> 32   uuid_32        54     278  24410 100
#> 33   uuid_33        52     306  40282 100
#> 34   uuid_34        46     284  43939 100
#> 35   uuid_35        49     242  25714 100
#> 36   uuid_36        47     424  22800 100
#> 37   uuid_37        48     313  45049 100
#> 38   uuid_38        54     313  29390 100
#> 39   uuid_39        52     407  39114 100
#> 40   uuid_40        54     231  44239 100
#> 41   uuid_41        53     368  27622 100
#> 42   uuid_42        45     356  23149 100
#> 43   uuid_43        47     406  44685 100
#> 44   uuid_44        47     500  24180 100
#> 45   uuid_45        52     406  28536 100
#> 46   uuid_46        47     387  33388 100
#> 47   uuid_47        51     446  43408 100
#> 48   uuid_48        46     429  29870 100
#> 49   uuid_49        45     362  25106 100
#> 50   uuid_50        50     431  26315 100
#> 51   uuid_51        55     487  30008 100
#> 52   uuid_52        50     297  43876 100
#> 53   uuid_53        50     218  35443 100
#> 54   uuid_54        46     449  26932 100
#> 55   uuid_55        48     272  33041 100
#> 56   uuid_56        49     223  37800 100
#> 57   uuid_57        54     281  30312 100
#> 58   uuid_58        54     292  29434 100
#> 59   uuid_59        51     208  28580 100
#> 60   uuid_60        52     203  46530 100
#> 61   uuid_61        54     435  45080 100
#> 62   uuid_62        54     233  46912 100
#> 63   uuid_63        49     252  47725 100
#> 64   uuid_64        50     352  37142 100
#> 65   uuid_65        49     325  34155 100
#> 66   uuid_66        53     403  24394 100
#> 67   uuid_67        49     379  20550 100
#> 68   uuid_68        47     379  40836 100
#> 69   uuid_69        46     221  29642 100
#> 70   uuid_70        54     236  37331 100
#> 71   uuid_71        49     288  46792 100
#> 72   uuid_72        46     206  45951 100
#> 73   uuid_73        49     360  34914 100
#> 74   uuid_74        52     289  21594 100
#> 75   uuid_75        49     232  37130 100
#> 76   uuid_76        48     375  39178 100
#> 77   uuid_77        51     341  42526 100
#> 78   uuid_78        46     401  28933 100
#> 79   uuid_79        49     394  31285 100
#> 80   uuid_80        54     402  31950 100
#> 81   uuid_81        47     341  47357 100
#> 82   uuid_82        53     260  49028 100
#> 83   uuid_83        51     407  21786 100
#> 84   uuid_84        46     444  48895 100
#> 85   uuid_85        45     392  45944 100
#> 86   uuid_86        52     209  46350 100
#> 87   uuid_87        52     357  21455 100
#> 88   uuid_88        52     295  31155 100
#> 89   uuid_89        46     341  44848 100
#> 90   uuid_90        52     497  49751 100
#> 91   uuid_91         1     217  40934 100
#> 92   uuid_92         1     200  44788 100
#> 93   uuid_93         0     254  45772 100
#> 94   uuid_94         1     347  30948 100
#> 95   uuid_95         0     205  46932 100
#> 96   uuid_96        99     600  26225 100
#> 97   uuid_97       100     100  28847 100
#> 98   uuid_98       100      80  45183 100
#> 99   uuid_99       100    1020  28226 100
#> 100 uuid_100        99    1050  41026  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         1
#>  5 uuid_95  outlier (normal distribution) one_value         0
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value       100
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>