check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_seperator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_seperator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        49     367     60 100
#> 2     uuid_2        46     470      0 100
#> 3     uuid_3        54     301     80 100
#> 4     uuid_4        46     236   1020 100
#> 5     uuid_5        53     406   1050 100
#> 6     uuid_6        48     413  48509 100
#> 7     uuid_7        48     493  46991 100
#> 8     uuid_8        55     377  41225 100
#> 9     uuid_9        48     287  35552 100
#> 10   uuid_10        48     413  20617 100
#> 11   uuid_11        46     437  46839 100
#> 12   uuid_12        53     323  39315 100
#> 13   uuid_13        46     497  33517 100
#> 14   uuid_14        46     490  30455 100
#> 15   uuid_15        48     294  32184 100
#> 16   uuid_16        51     496  33140 100
#> 17   uuid_17        54     427  49222 100
#> 18   uuid_18        48     202  46040 100
#> 19   uuid_19        50     320  36390 100
#> 20   uuid_20        52     299  27712 100
#> 21   uuid_21        52     223  26193 100
#> 22   uuid_22        47     331  44911 100
#> 23   uuid_23        46     236  49115 100
#> 24   uuid_24        48     430  21277 100
#> 25   uuid_25        50     346  33153 100
#> 26   uuid_26        46     427  28224 100
#> 27   uuid_27        52     419  48527 100
#> 28   uuid_28        50     286  40667 100
#> 29   uuid_29        48     272  20237 100
#> 30   uuid_30        51     351  29044 100
#> 31   uuid_31        54     492  32342 100
#> 32   uuid_32        46     281  20099 100
#> 33   uuid_33        52     209  26406 100
#> 34   uuid_34        45     465  34555 100
#> 35   uuid_35        51     290  27078 100
#> 36   uuid_36        53     455  42993 100
#> 37   uuid_37        54     452  45511 100
#> 38   uuid_38        54     439  37587 100
#> 39   uuid_39        45     458  42876 100
#> 40   uuid_40        46     390  45648 100
#> 41   uuid_41        47     385  46770 100
#> 42   uuid_42        46     493  31728 100
#> 43   uuid_43        51     441  33354 100
#> 44   uuid_44        47     332  21390 100
#> 45   uuid_45        55     361  33404 100
#> 46   uuid_46        52     232  33588 100
#> 47   uuid_47        48     251  45940 100
#> 48   uuid_48        54     492  42589 100
#> 49   uuid_49        46     410  22577 100
#> 50   uuid_50        54     434  20433 100
#> 51   uuid_51        50     202  38003 100
#> 52   uuid_52        54     325  34400 100
#> 53   uuid_53        51     374  38074 100
#> 54   uuid_54        46     364  22634 100
#> 55   uuid_55        46     234  32082 100
#> 56   uuid_56        49     269  20034 100
#> 57   uuid_57        54     268  46492 100
#> 58   uuid_58        48     356  39080 100
#> 59   uuid_59        47     225  41982 100
#> 60   uuid_60        47     418  26764 100
#> 61   uuid_61        55     496  27324 100
#> 62   uuid_62        51     235  34624 100
#> 63   uuid_63        47     275  47830 100
#> 64   uuid_64        50     369  45417 100
#> 65   uuid_65        52     467  25316 100
#> 66   uuid_66        50     367  31682 100
#> 67   uuid_67        47     278  20341 100
#> 68   uuid_68        50     450  26498 100
#> 69   uuid_69        47     289  39471 100
#> 70   uuid_70        53     207  23102 100
#> 71   uuid_71        52     398  21338 100
#> 72   uuid_72        46     332  36558 100
#> 73   uuid_73        46     355  28424 100
#> 74   uuid_74        53     446  38045 100
#> 75   uuid_75        53     428  33386 100
#> 76   uuid_76        50     345  24573 100
#> 77   uuid_77        49     211  34953 100
#> 78   uuid_78        50     287  28663 100
#> 79   uuid_79        53     410  38421 100
#> 80   uuid_80        52     445  35496 100
#> 81   uuid_81        50     412  22424 100
#> 82   uuid_82        54     342  47789 100
#> 83   uuid_83        47     453  47486 100
#> 84   uuid_84        53     340  28940 100
#> 85   uuid_85        46     271  33352 100
#> 86   uuid_86        54     391  27840 100
#> 87   uuid_87        54     380  32536 100
#> 88   uuid_88        50     327  25677 100
#> 89   uuid_89        51     223  22988 100
#> 90   uuid_90        48     334  41347 100
#> 91   uuid_91         0     353  39736 100
#> 92   uuid_92         1     214  34469 100
#> 93   uuid_93         1     401  26598 100
#> 94   uuid_94         0     298  27547 100
#> 95   uuid_95         1     201  36979 100
#> 96   uuid_96       100     600  44029 100
#> 97   uuid_97       100     100  45683 100
#> 98   uuid_98        99      80  42952 100
#> 99   uuid_99       100    1020  22332 100
#> 100 uuid_100        99    1050  39722  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         0
#>  2 uuid_92  outlier (normal distribution) one_value         1
#>  3 uuid_93  outlier (normal distribution) one_value         1
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value       100
#>  7 uuid_97  outlier (normal distribution) one_value       100
#>  8 uuid_98  outlier (normal distribution) one_value        99
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value        99
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>