check outliers over the dataset

check outliers over the dataset

Usage

check_outliers(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_separator = ".",
  columns_not_to_check = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

uuid_column

UUID. Default is uuid

element_name

name of the dataset in list

kobo_survey

Kobo survey sheet. Default is NULL.

kobo_choices

Kobo choices sheet. Default is NULL.

cols_to_add_cleaning_log

Variables those must be included in the output

strongness_factor

Strongness factor define how strong your outliers will be. The default is 3.

minimum_unique_value_of_variable

Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.

remove_choice_multiple

TRUE (default) will remove choice multiple questions from the output.

sm_separator

Separator for choice multiple questions. The default is "."

columns_not_to_check

Columns to exclude from the checks even if they are numeric values.

Value

return a list with the dataset checked stored as checked_dataset and a dataframe with the outliers log

Examples

dataset_outlier <- data.frame(
  uuid = paste0("uuid_", 1:100),
  one_value = c(round(runif(90, min = 45, max = 55)), round(runif(5)), round(runif(5, 99, 100))),
  expense = c(sample(200:500, replace = TRUE, size = 95), c(600, 100, 80, 1020, 1050)),
  income = c(c(60, 0, 80, 1020, 1050), sample(20000:50000, replace = TRUE, size = 95)),
  yy = c(rep(100, 99), 10)
)

check_outliers(dataset = dataset_outlier, uuid_column = "uuid")
#> [1] "checking_one_value"
#> [1] "checking_expense"
#> [1] "checking_income"
#> [1] "checking_yy"
#> $checked_dataset
#>         uuid one_value expense income  yy
#> 1     uuid_1        46     212     60 100
#> 2     uuid_2        48     384      0 100
#> 3     uuid_3        47     407     80 100
#> 4     uuid_4        48     469   1020 100
#> 5     uuid_5        51     238   1050 100
#> 6     uuid_6        47     385  32276 100
#> 7     uuid_7        55     215  46287 100
#> 8     uuid_8        53     230  31797 100
#> 9     uuid_9        46     386  40249 100
#> 10   uuid_10        52     469  26447 100
#> 11   uuid_11        53     434  34260 100
#> 12   uuid_12        50     324  46649 100
#> 13   uuid_13        54     486  47931 100
#> 14   uuid_14        52     456  49936 100
#> 15   uuid_15        55     252  33353 100
#> 16   uuid_16        48     284  20070 100
#> 17   uuid_17        51     411  26348 100
#> 18   uuid_18        48     229  48989 100
#> 19   uuid_19        49     377  36744 100
#> 20   uuid_20        47     412  49944 100
#> 21   uuid_21        55     416  41775 100
#> 22   uuid_22        49     201  34981 100
#> 23   uuid_23        49     350  48149 100
#> 24   uuid_24        52     275  31617 100
#> 25   uuid_25        51     280  35160 100
#> 26   uuid_26        54     280  47355 100
#> 27   uuid_27        54     211  30348 100
#> 28   uuid_28        46     324  33409 100
#> 29   uuid_29        52     420  43316 100
#> 30   uuid_30        45     340  27674 100
#> 31   uuid_31        50     298  24670 100
#> 32   uuid_32        55     476  39278 100
#> 33   uuid_33        54     396  48112 100
#> 34   uuid_34        48     304  33079 100
#> 35   uuid_35        54     274  21374 100
#> 36   uuid_36        53     500  20490 100
#> 37   uuid_37        51     477  35246 100
#> 38   uuid_38        51     492  45714 100
#> 39   uuid_39        52     398  29215 100
#> 40   uuid_40        49     455  20173 100
#> 41   uuid_41        51     327  40942 100
#> 42   uuid_42        54     495  23722 100
#> 43   uuid_43        51     228  46754 100
#> 44   uuid_44        54     309  25700 100
#> 45   uuid_45        53     276  46578 100
#> 46   uuid_46        51     430  25198 100
#> 47   uuid_47        47     391  39097 100
#> 48   uuid_48        49     286  37436 100
#> 49   uuid_49        49     261  31229 100
#> 50   uuid_50        52     249  46757 100
#> 51   uuid_51        54     370  27342 100
#> 52   uuid_52        54     325  49387 100
#> 53   uuid_53        45     209  36718 100
#> 54   uuid_54        52     372  44819 100
#> 55   uuid_55        48     301  40387 100
#> 56   uuid_56        48     494  36870 100
#> 57   uuid_57        47     337  24636 100
#> 58   uuid_58        50     210  23674 100
#> 59   uuid_59        54     328  49929 100
#> 60   uuid_60        47     442  25555 100
#> 61   uuid_61        48     229  33583 100
#> 62   uuid_62        55     282  33631 100
#> 63   uuid_63        48     273  34612 100
#> 64   uuid_64        52     291  32957 100
#> 65   uuid_65        50     201  46131 100
#> 66   uuid_66        53     236  45459 100
#> 67   uuid_67        48     367  33800 100
#> 68   uuid_68        48     245  36984 100
#> 69   uuid_69        50     497  25940 100
#> 70   uuid_70        54     441  41860 100
#> 71   uuid_71        53     400  39238 100
#> 72   uuid_72        53     473  37080 100
#> 73   uuid_73        46     297  41120 100
#> 74   uuid_74        51     347  24012 100
#> 75   uuid_75        55     290  32181 100
#> 76   uuid_76        54     465  43416 100
#> 77   uuid_77        49     322  33824 100
#> 78   uuid_78        50     307  39921 100
#> 79   uuid_79        51     205  33400 100
#> 80   uuid_80        48     334  39623 100
#> 81   uuid_81        53     478  49174 100
#> 82   uuid_82        47     301  45773 100
#> 83   uuid_83        47     390  36122 100
#> 84   uuid_84        51     220  44801 100
#> 85   uuid_85        47     419  45365 100
#> 86   uuid_86        46     335  38585 100
#> 87   uuid_87        52     418  27284 100
#> 88   uuid_88        54     304  21978 100
#> 89   uuid_89        51     298  25261 100
#> 90   uuid_90        55     239  28145 100
#> 91   uuid_91         1     288  30162 100
#> 92   uuid_92         0     262  26046 100
#> 93   uuid_93         0     332  28389 100
#> 94   uuid_94         0     388  26167 100
#> 95   uuid_95         1     424  41240 100
#> 96   uuid_96        99     600  36594 100
#> 97   uuid_97        99     100  27883 100
#> 98   uuid_98       100      80  42443 100
#> 99   uuid_99       100    1020  44642 100
#> 100 uuid_100       100    1050  46624  10
#> 
#> $potential_outliers
#> # A tibble: 18 × 4
#>    uuid     issue                         question  old_value
#>    <chr>    <chr>                         <chr>         <dbl>
#>  1 uuid_91  outlier (normal distribution) one_value         1
#>  2 uuid_92  outlier (normal distribution) one_value         0
#>  3 uuid_93  outlier (normal distribution) one_value         0
#>  4 uuid_94  outlier (normal distribution) one_value         0
#>  5 uuid_95  outlier (normal distribution) one_value         1
#>  6 uuid_96  outlier (normal distribution) one_value        99
#>  7 uuid_97  outlier (normal distribution) one_value        99
#>  8 uuid_98  outlier (normal distribution) one_value       100
#>  9 uuid_99  outlier (normal distribution) one_value       100
#> 10 uuid_100 outlier (normal distribution) one_value       100
#> 11 uuid_99  outlier (normal distribution) expense        1020
#> 12 uuid_100 outlier (normal distribution) expense        1050
#> 13 uuid_97  outlier (log distribution)    expense         100
#> 14 uuid_98  outlier (log distribution)    expense          80
#> 15 uuid_1   outlier (log distribution)    income           60
#> 16 uuid_2   outlier (log distribution)    income            0
#> 17 uuid_3   outlier (log distribution)    income           80
#> 18 uuid_100 outlier (normal distribution) yy               10
#>