Checks for potential PII

Checks for potential PII

Usage

check_pii(
  dataset,
  element_name = "checked_dataset",
  uuid_column = "uuid",
  words_to_look = NULL
)

Arguments

dataset

dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"

element_name

If the input is a list file, please specify the element name that contains the dataset

uuid_column

uuid column in the dataset. Default is uuid.

words_to_look

Specify the words that might be the PIIs

Value

dataset with potential PII

Examples

check_pii(cleaningtools_raw_data, uuid_column = "X_uuid")
#> $checked_dataset
#> # A tibble: 632 × 348
#>    X.U.FEFF.start    end   date_assessment deviceid enumerator_num neighbourhood
#>    <chr>             <chr> <chr>           <chr>             <int> <chr>        
#>  1 2021-07-05T10:58… 2021… 2021-07-05      collect…             13 A2           
#>  2 2021-07-05T10:08… 2021… 2021-07-05      collect…             13 A2           
#>  3 2021-07-05T11:22… 2021… 2021-07-05      collect…              8 A2           
#>  4 2021-07-04T22:22… 2021… 2021-07-04      collect…             15 A2           
#>  5 2021-07-04T23:11… 2021… 2021-07-04      collect…             15 A2           
#>  6 2021-07-05T10:17… 2021… 2021-07-05      collect…             14 A2           
#>  7 2021-07-05T11:38… 2021… 2021-07-05      collect…              4 A4           
#>  8 2021-07-05T11:01… 2021… 2021-07-05      collect…             16 A2           
#>  9 2021-07-05T11:45… 2021… 2021-07-05      collect…             16 A2           
#> 10 2021-07-05T10:19… 2021… 2021-07-05      collect…              8 A2           
#> # ℹ 622 more rows
#> # ℹ 342 more variables: consent_remote <chr>, hhh_r <chr>,
#> #   will_to_response_r <chr>, age_respondent_r <int>, age_hoh <int>,
#> #   gender_hoh <chr>, displace_status <chr>, displace_status_returnee <chr>,
#> #   return_date <chr>, num_hh_member <int>, hh_hosting <chr>,
#> #   num_hh_hosting <int>, hh_hosted <chr>, shelter_occupation <chr>,
#> #   shelter_occupation_other <chr>, property_title <chr>, …
#> 
#> $potential_PII
#> # A tibble: 5 × 3
#>   uuid  question                              issue        
#>   <chr> <chr>                                 <chr>        
#> 1 all   neighbourhood                         Potential PII
#> 2 all   water_supply_rest_neighbourhood       Potential PII
#> 3 all   water_supply_other_neighbourhoods     Potential PII
#> 4 all   water_supply_other_neighbourhoods_why Potential PII
#> 5 all   consent_telephone_number              Potential PII
#>